+++ packages/core/components/login-screen/login-screen.js
@@ -17,11 +17,11 @@ export default {
   clicks: {
     '.login-screen-open': function openLoginScreen($clickedEl, data = {}) {
       const app = this;
-      app.loginScreen.open(data.loginScreen, data.animate);
+      app.loginScreen.open(data.loginScreen, data.animate, $clickedEl);
     },
     '.login-screen-close': function closeLoginScreen($clickedEl, data = {}) {
       const app = this;
-      app.loginScreen.close(data.loginScreen, data.animate);
+      app.loginScreen.close(data.loginScreen, data.animate, $clickedEl);
     },
   },
 };