+++ packages/vue/components/list-input.js
@@ -56,6 +56,7 @@ export default {
     pattern: String,
     validate: [Boolean, String],
     validateOnBlur: Boolean,
+    onValidate: Function,
     tabindex: [String, Number],
     resizable: Boolean,
     clearButton: Boolean,
@@ -495,14 +496,20 @@ export default {
       if (!f7 || !inputEl) return;
       const validity = inputEl.validity;
       if (!validity) return;
+      const {
+        onValidate
+      } = self.props;
 
       if (!validity.valid) {
+        if (onValidate) onValidate(false);
+
         if (self.state.inputInvalid !== true) {
           self.setState({
             inputInvalid: true
           });
         }
       } else if (self.state.inputInvalid !== false) {
+        if (onValidate) onValidate(true);
         self.setState({
           inputInvalid: false
         });